refactor: 사용 방식에 용도에 따라 constant 분리
