+++ src/main/java/step3/constant/LottoViewConstant.java
@@ -1,0 +1,7 @@
+package step3.constant;
+
+public class LottoViewConstant {
+    public static String OVERVIEW_FORMAT = "%s개 일치 (%s원)- %s개";
+    public static String REWARD_RATE_FORMAT = "총 수익률은 %.2f입니다.(기준이 1이기 때문에 결과적으로 %s라는 의미임)";
+    public static String LOTTOS_INFO_FORMAT = "%s개를 구매했습니다.";
+}
